Add compile_model and StanModel.from_stan_file
