+++ python/test/data/syntax_error.stan
@@ -1,0 +1,6 @@
+parameters {
+   real theta;
+}
+model {
+   theta ~ normal()
+}
