+++ python/test/test_compile.py
@@ -1,0 +1,48 @@
+from pathlib import Path
+
+import bridgestan as bs
+import pytest
+
+HERE = Path(__file__).parent
+STAN_FOLDER = HERE.parent.parent / "stan"
+
+def test_compile_good():
+    stanfile = STAN_FOLDER / "multi" / "multi.stan"
+    lib = bs.compile.generate_so_name(stanfile)
+    lib.unlink(missing_ok=True)
+    res = bs.compile_model(stanfile)
+    assert lib.samefile(res)
+    lib.unlink()
+    res = bs.compile_model(stanfile, args=["STAN_THREADS=true"])
+    assert lib.samefile(res)
+
+def test_compile_bad_ext():
+    not_stanfile = STAN_FOLDER / "multi" / "multi.data.json"
+    with pytest.raises(ValueError, match=r".stan"):
+        res = bs.compile_model(not_stanfile)
+
+def test_compile_nonexistant():
+    not_stanfile = STAN_FOLDER / "multi" / "multi-nothere.data.json"
+    with pytest.raises(FileNotFoundError):
+        res = bs.compile_model(not_stanfile)
+
+def test_compile_syntax_error():
+    stanfile = HERE / "data" / "syntax_error.stan"
+    with pytest.raises(RuntimeError, match=r"Syntax error"):
+        bs.compile_model(stanfile)
+
+def test_compile_bad_cmdstan():
+    stanfile = STAN_FOLDER / "multi" / "multi.stan"
+    old_path = bs.compile.CMDSTAN_PATH
+    bs.compile.set_cmdstan_path("dummy")
+    with pytest.raises(RuntimeError, match=r"Make sure CmdStan is installed"):
+        bs.compile_model(stanfile)
+    bs.compile.set_cmdstan_path(old_path)
+
+def test_compile_bad_bridgestan():
+    old_path = bs.compile.BRIDGESTAN_PATH
+    with pytest.raises(ValueError, match=r"does not exist"):
+        bs.compile.set_bridgestan_path('dummy')
+    with pytest.raises(ValueError, match=r"does not contain file 'Makefile'"):
+        bs.compile.set_bridgestan_path(str(HERE))
+    bs.compile.set_bridgestan_path(old_path)
